Small refactor, migrations for tests
