+++ src/test/java/com/wine/to/up/user/service/controller/AuthenticationControllerIntegrationTest.java
@@ -2,6 +2,8 @@ package com.wine.to.up.user.service.controller;
 
 import static org.junit.Assert.assertEquals;
 
+import com.wine.to.up.user.service.domain.dto.RoleDto;
+import com.wine.to.up.user.service.domain.dto.UserDto;
 import com.wine.to.up.user.service.security.JwtTokenProvider;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -14,7 +16,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @SpringBootTest
 @RunWith(SpringRunner.class)
-@TestPropertySource("classpath:test.properties")
+@TestPropertySource("classpath:application.properties")
 public class AuthenticationControllerIntegrationTest {
     @Autowired
     private AuthenticationController authenticationController;
@@ -24,7 +26,13 @@ public class AuthenticationControllerIntegrationTest {
 
     @Test
     public void testValidateCorrect() {
-        String token = jwtTokenProvider.createToken("+71234567890", true);
+        String token = jwtTokenProvider.createToken(
+            new UserDto()
+                .setId(1L)
+                .setPhoneNumber("+71234567890")
+                .setRole(new RoleDto().setId(1L).setName("test")),
+            true
+        );
 
         ResponseEntity<Void> response = this.authenticationController.validate(token);
         assertEquals(HttpStatus.OK, response.getStatusCode());
