+++ src/test/resources/application.properties
@@ -6,9 +6,19 @@ jdbc.url=jdbc:h2:mem:myDb
 
 spring.flyway.url = jdbc:h2:mem:myDb
 spring.flyway.baseline-on-migrate = true
+spring.flyway.locations = classpath:db/test-migration
 
 hibernate.dialect=org.hibernate.dialect.H2Dialect
 hibernate.hbm2ddl.auto=validate
 
 default.kafka.bootstrap-server=localhost:29092
 spring.kafka.bootstrap-server=${S_KAFKA_BOOTSTRAP_HOST:${default.kafka.bootstrap-server}}
+
+jwt.token.secret=jwtsecret
+jwt.token.expired.access=3600000
+jwt.token.expired.refresh=2592000000
+
+user.role.user.id=1
+user.role.admin.id=2
+
+firebase.url=https://testfire-1bc2f.firebaseio.com
