+++ src/test/resources/db/test-migration/V1.0.1__default_roles.sql
@@ -1,0 +1,5 @@
+INSERT INTO
+    user_roles ("id", "name")
+VALUES
+    (1, 'USER'),
+    (2, 'ADMIN');
